feat: finish componentization
