+++ lib/conversion/controller/validate_provider.dart
@@ -1,0 +1,3 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final validateProvider = StateProvider((ref) => false);
