+++ lib/conversion/widgets/floating_cripto_button.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:projeto_criptos/conversion/controller/validate_provider.dart';
+
+import '../../shared/templates/app_assets.dart';
+
+class FloatingCriptoButton extends HookConsumerWidget {
+  const FloatingCriptoButton({
+    Key? key,
+  }) : super(key: key);
+
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    bool validate = ref.watch(validateProvider.state).state;
+    return FloatingActionButton(
+      backgroundColor: validate ? AppAssets.magenta : Colors.grey,
+      onPressed: validate ? () {} : null,
+      child: const Icon(
+        Icons.keyboard_arrow_right,
+      ),
+    );
+  }
+}
